+++ MedImgPkg/include/arithmetic/mi_run_length_operator.h
@@ -10,7 +10,7 @@ class Arithmetic_Export RunLengthOperator {
 public:
     static std::vector<unsigned int> encode(const std::vector<unsigned char>& to_be_encoded);
     static std::vector<unsigned int> encode(const unsigned char* mask_array_pointer,
-            const size_t total_number_of_voxels);
+            const unsigned int total_number_of_voxels);
 
     static std::vector<unsigned char> decode(const std::vector<unsigned int>& to_be_decoded);
     static int decode(unsigned int* code_buffer , unsigned int code_buffer_len ,
